@@ -1,6 +1,7 @@
-import 'dart:async';
+/*import 'dart:async';
 
 import 'package:flutter/material.dart';
+//import 'package:webview_flutter/webview_flutter.dart';
 import 'package:flutter_webview_plugin/flutter_webview_plugin.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:uuid/uuid.dart';
@@ -213,3 +214,4 @@ class _MyWebViewState extends State<MyWebView> {
     print("URL " + selectedUrl);
   }
 }
+*/
