@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:flutter_background_geolocation/flutter_background_geolocation.dart' as bg;
+import 'package:flutter_background_geolocation/flutter_background_geolocation.dart'
+    as bg;
 import 'ENV.dart';
 
 void _onHttp(bg.HttpEvent event) async {
-
   switch (event.status) {
     case 403:
     case 406:
@@ -19,7 +19,8 @@ void _onHttp(bg.HttpEvent event) async {
       }
       break;
     case 410:
-      print('[TransistorAuth] It seems this device has been destroyed from the server.  The authentication token is no longer valid.');
+      print(
+          '[TransistorAuth] It seems this device has been destroyed from the server.  The authentication token is no longer valid.');
       // For now in this case we will do nothing. Consider alternatives
       break;
   }
@@ -32,18 +33,19 @@ class TransistorAuth {
     try {
       SharedPreferences prefs = await _prefs;
       // Request a JWT from server
-      String sample_id = prefs.getString("sample_id");
-      String user_uuid = prefs.getString("user_uuid");
+      String? sample_id = prefs.getString("sample_id");
+      String? user_uuid = prefs.getString("user_uuid");
       if (sample_id == null || user_uuid == null) {
         // TODO throw an Error instead.
         return false;
       }
 
-      bg.TransistorAuthorizationToken jwt = await bg.TransistorAuthorizationToken.findOrCreate(sample_id, user_uuid, ENV.TRACKER_HOST);
+      bg.TransistorAuthorizationToken jwt =
+          await bg.TransistorAuthorizationToken.findOrCreate(
+              sample_id, user_uuid, ENV.TRACKER_HOST);
 
-      await bg.BackgroundGeolocation.setConfig(bg.Config(
-          transistorAuthorizationToken: jwt
-      ));
+      await bg.BackgroundGeolocation.setConfig(
+          bg.Config(transistorAuthorizationToken: jwt));
       return true;
     } catch (error) {
       print("[ERROR] $error");
@@ -54,7 +56,7 @@ class TransistorAuth {
 
   static Future<void> registerErrorHandler() async {
     bg.State state = await bg.BackgroundGeolocation.state;
-    if ((state.params != null) && (state.params['device'] != null)) {
+    if ((state.params != null) && (state.params!['device'] != null)) {
       _migrateConfig();
     }
     bg.BackgroundGeolocation.removeListener(_onHttp);
@@ -72,7 +74,6 @@ class TransistorAuth {
         stopOnTerminate: false,
         startOnBoot: true,
         url: "${ENV.TRACKER_HOST}/api/locations",
-        params: {}
-    ));
+        params: {}));
   }
 }
